- fix analysis
