+++ per_data/005930.csv
@@ -1,2 +1,2 @@
 ﻿,company_name,company_code,PER,동일업종 PER
-0,삼성전자,005930,21.66,19.33
+0,삼성전자,005930,21.50,18.78
