+++ rate/000270.csv
@@ -1,2 +1,2 @@
 ratio,portion
-0.43553270956185264,-0.4
+0.43553270956185264,0.049340187131444196
