+++ rate/000660.csv
@@ -1,2 +1,2 @@
 ratio,portion
-0.7863270648950533,0.4
+0.7863270648950533,-0.055898119468516
