+++ rate/003550.csv
@@ -1,2 +1,2 @@
 ratio,portion
-0.6434469847680021,0.1
+0.6434469847680021,-0.013034095430400638
