+++ rate/005380.csv
@@ -1,2 +1,2 @@
 ratio,portion
-0.6186916963230845,0.0
+0.6186916963230845,-0.005607508896925362
