+++ rate/005930.csv
@@ -1,2 +1,2 @@
 ratio,portion
-0.7080388576532931,0.2
+0.5510923746385751,0.03979693328033029
