+++ rate/008770.csv
@@ -1,2 +1,2 @@
 ratio,portion
-0.8676862713670446,0.6000000000000001
+0.8676862713670446,-0.08030588141011337
