+++ rate/009150.csv
@@ -1,2 +1,2 @@
 ratio,portion
-0.6470202162534432,0.1
+0.6470202162534432,-0.014106064876032965
