+++ rate/030200.csv
@@ -1,2 +1,2 @@
 ratio,portion
-0.3946940282192247,-0.5
+0.3946940282192247,0.061591791534232575
