+++ rate/034730.csv
@@ -1,2 +1,2 @@
 ratio,portion
-0.7100084819284015,0.2
+0.7100084819284015,-0.033002544578520455
