+++ rate/035420.csv
@@ -1,2 +1,2 @@
 ratio,portion
-0.7624218986773187,0.4
+0.7624218986773187,-0.04872656960319561
