+++ rate/035720.csv
@@ -1,2 +1,2 @@
 ratio,portion
-0.6120992295903452,0.0
+0.6120992295903452,-0.0036297688771035784
