+++ rate/051900.csv
@@ -1,2 +1,2 @@
 ratio,portion
-0.7167444923891367,0.2
+0.7167444923891367,-0.035023347716741024
