+++ rate/051910.csv
@@ -1,2 +1,2 @@
 ratio,portion
-0.6776826157516641,0.1
+0.6776826157516641,-0.02330478472549924
