+++ rate/068270.csv
@@ -1,2 +1,2 @@
 ratio,portion
-0.937637570884693,0.8
+0.937637570884693,-0.1012912712654079
