+++ rate/139480.csv
@@ -1,2 +1,2 @@
 ratio,portion
-0.5345731788883339,-0.1
+0.5345731788883339,0.019628046333499828
