+++ rate/251270.csv
@@ -1,2 +1,2 @@
 ratio,portion
-0.8435685306599051,0.6000000000000001
+0.8435685306599051,-0.07307055919797154
