+++ tmp/checkpoint
@@ -1,0 +1,2 @@
+model_checkpoint_path: "ckeckpointer.ckpt"
+all_model_checkpoint_paths: "ckeckpointer.ckpt"
